RD-5020 Split `blueprints_spec` into 2 files (#2099)
* Split `blueprints_spec` into 2 files
* Refactor 2 tests in blueprints_spec
* Use local stubbing for window in blueprints_spec
